Fixed Issue with Leader Leaving
When the group leader leaves the group, if there are other members in the group, it will pick one and make them the leader, otherwise it will destroy the group.  It will now do the same thing when the leader quits and properly removes members when they leave or quit.